move policy to act folder
